requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.6.1
pyyaml==6.0.1
aiofiles==23.2.1
python-telegram-bot==21.0
pdf2image==1.16.3
# PaddleOCR for text extraction
paddlepaddle>=2.6.2
paddleocr>=2.7.0
prometheus-fastapi-instrumentator>=6.1.0
